+++ python/ray/autoscaler/_private/aws/config.py
@@ -11,7 +11,7 @@ import boto3
 from botocore.config import Config
 import botocore
 
-from ray.autoscaler.constants import BOTO_MAX_RETRIES
+from ray.autoscaler._private.constants import BOTO_MAX_RETRIES
 from ray.autoscaler.tags import NODE_KIND_WORKER, NODE_KIND_HEAD
 from ray.autoscaler._private.providers import _PROVIDER_PRETTY_NAMES
 from ray.autoscaler._private.aws.utils import LazyDefaultDict, \
